refactor: move shortcuts to legacy package
