+++ app/src/main/kotlin/io/homeassistant/companion/android/settings/SettingsFragment.kt
@@ -48,7 +48,7 @@ import io.homeassistant.companion.android.settings.qs.ManageTilesFragment
 import io.homeassistant.companion.android.settings.sensor.SensorSettingsFragment
 import io.homeassistant.companion.android.settings.sensor.SensorUpdateFrequencyFragment
 import io.homeassistant.companion.android.settings.server.ServerSettingsFragment
-import io.homeassistant.companion.android.settings.shortcuts.ManageShortcutsSettingsFragment
+import io.homeassistant.companion.android.settings.shortcuts.legacy.ManageShortcutsSettingsFragment
 import io.homeassistant.companion.android.settings.vehicle.ManageAndroidAutoSettingsFragment
 import io.homeassistant.companion.android.settings.wear.SettingsWearActivity
 import io.homeassistant.companion.android.settings.wear.SettingsWearDetection
@@ -216,7 +216,7 @@ class SettingsFragment(
                 findPreference<Preference>("manage_shortcuts")?.setOnPreferenceClickListener {
                     parentFragmentManager.commit {
                         replace(R.id.content, ManageShortcutsSettingsFragment::class.java, null)
-                        addToBackStack(getString(commonR.string.shortcuts))
+                        addToBackStack(getString(commonR.string.shortcuts_legacy))
                     }
                     return@setOnPreferenceClickListener true
                 }
