+++ app/src/main/kotlin/io/homeassistant/companion/android/settings/shortcuts/legacy/views/ManageShortcutsView.kt
@@ -1,4 +1,4 @@
-package io.homeassistant.companion.android.settings.shortcuts.views
+package io.homeassistant.companion.android.settings.shortcuts.legacy.views
 
 import android.os.Build
 import android.widget.Toast
@@ -39,8 +39,8 @@ import androidx.compose.ui.unit.sp
 import com.mikepenz.iconics.compose.IconicsPainter
 import io.homeassistant.companion.android.common.R
 import io.homeassistant.companion.android.common.compose.theme.HATheme
-import io.homeassistant.companion.android.settings.shortcuts.ManageShortcutsSettingsFragment
-import io.homeassistant.companion.android.settings.shortcuts.ManageShortcutsViewModel
+import io.homeassistant.companion.android.settings.shortcuts.legacy.ManageShortcutsSettingsFragment
+import io.homeassistant.companion.android.settings.shortcuts.legacy.ManageShortcutsViewModel
 import io.homeassistant.companion.android.util.compose.ServerExposedDropdownMenu
 import io.homeassistant.companion.android.util.compose.entity.EntityPicker
 import io.homeassistant.companion.android.util.plus
