+++ common/src/main/res/values/strings.xml
@@ -826,6 +826,7 @@
     <string name="shortcuts_tile_log_in">Log in to add your first shortcut</string>
     <string name="shortcuts_tile_no_tiles_yet">There are no shortcuts tiles added yet</string>
     <string name="shortcuts">Shortcuts</string>
+    <string name="shortcuts_legacy">Shortcuts (Legacy)</string>
     <string name="show">Show</string>
     <string name="show_on_map">Show on map</string>
     <string name="show_share_logs_summary">Sharing logs with the Home Assistant team will help to solve issues. Please share the logs only if you have been asked to do so by a Home Assistant developer</string>
